requirements.txt
# Core Dependencies
streamlit==1.28.0
tensorflow==2.15.0
flask==3.0.0

# Data Processing
numpy==1.24.3
pandas==2.1.1
Pillow==10.1.0
scikit-learn==1.3.1

# API and Web
requests==2.31.0
werkzeug==3.0.1

# Database (SQLite is built-in, but we need mysql-connector for compatibility)
# Note: We'll use SQLite locally but keep the structure compatible

# Utilities
python-dotenv==1.0.0

# Optional but recommended
matplotlib==3.8.0
seaborn==0.13.0